requirements.txt
black==22.10.0
jupyter==1.0.0
matplotlib==3.7.1
numpy==1.23.4
pandas==1.5.1
PyQt5==5.15.7
pylint==2.15.5
pytest==7.2.0
pytest-cov==4.0.0
pytest-xdist==3.0.2
Biopython==1.80
mypy==1.3.0
flake8==6.0.0
pandas==2.0.2
PyYAML==6.0
